# swagger.v1.yaml
openapi: 3.0.0
info:
  title: Banking Application
  version: '1.0'
  description: ''
servers:
  - url: 'http://localhost:3000/api/v1'
    description: Clients
paths:
...
  '/staff/analytics/receivable/:groupBy':
    get:
      summary: Get total money received
      tags:
        - Staff Analytics
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      anyOf:
                        - properties:
                            receivedAmount:
                              type: number
                            day:
                              type: string
                        - properties:
                            receivedAmount:
                              type: number
                            week:
                              type: string
                        - properties:
                            receivedAmount:
                              type: number
                            month:
                              type: string
                      type: object
      operationId: 'get-staff-analytics-receivable-:groupBy'
      security:
        - Authorization: []
        - Authorization: []
...
components:
...
  securitySchemes:
    Authorization:
      type: http
      scheme: bearer
security:
  - Authorization: []